api-gtw-acck/requirements.txt
aws-psycopg2==1.3.8
ulid==1.1